You can now add and remove images from an existing console. Update Console Management Page design
